+++ src/components/Widgets/Cart/ThankYouPage/ThankYouPage.vue
@@ -1,5 +1,8 @@
 <template>
   <div class="cart-container">
+    <div v-if="host">
+      ({{ host }})
+    </div>
     <template v-if="loading && !isEwanoUser">
       <q-skeleton type="circle" />
     </template>
@@ -48,6 +51,7 @@
 </template>
 
 <script>
+import { Capacitor } from '@capacitor/core'
 import { mixinAuth } from 'src/mixin/Mixins.js'
 import { APIGateway } from 'src/api/APIGateway.js'
 import mixinEwano from 'src/components/Widgets/Ewano/mixinEwano.js'
@@ -58,6 +62,7 @@ export default {
   mixins: [mixinAuth, mixinEwano],
   data () {
     return {
+      host: null,
       loading: false,
       hasPaid: false
     }
@@ -76,6 +81,17 @@ export default {
     this.$bus.on('ThankYouPageInvoiceLoading', (status) => {
       this.loading = status
     })
+    // alaa-app://alaatv.com/payment
+    if (Capacitor.isNativePlatform()) {
+      this.host = window.location.href
+    }
+    if (Capacitor.isNativePlatform() && window.location.href.indexOf('https://localhost') !== 0) {
+      const pathArray = window.location.href.split('/')
+      const orderId = pathArray[pathArray.length - 2]
+      this.$router.push({ name: 'UserPanel.ThankYouPage', params: { orderId } })
+      // document.location = window.location.href.replace('https://alaatv.com', 'alaa-app://alaatv.com')
+      //   .replace('http://alaatv.com', 'alaa-app://alaatv.com')
+    }
   },
   methods: {
     pushPurchaseEvent (order) {